Fixed the reversed traces.
The problem here is that Reconstruct.exe stores its points in reverse
order (last to first), and the Java version tries to imitate that by
doing the same. However, the handleing of the handles complicates this
process because the handles are effectively "between" the points.

Until this is worked out, it's important to be careful when opening and
saving files.