Moved the series files to a subdirectory indexed by sample number.
It's anticipated that various people (and tools) will produce traced
results for this project. These results can be collected here in the
"series/sample_#" subdirectories for review. This change simply moves
the existing series files into the "series/sample_1" subdirectory to
set an example for other samples that might be collected.